Move shared folders into shared folder
