+++ server/src/shared/constants/tableNames.ts
