+++ server/src/shared/database/index.ts
@@ -1,7 +1,7 @@
 import knex from 'knex';
 
 // eslint-disable-next-line @typescript-eslint/no-var-requires
-const knexfile = require('../../knexfile');
+const knexfile = require('../../../knexfile');
 
 const environment = process.env.ENVIRONMENT || 'development';
